+++ src/test/java/com/maxmind/minfraud/request/CreditCardTest.java
@@ -14,10 +14,10 @@ public class CreditCardTest {
     @Test
     public void testIssuerIdNumber() {
         CreditCard cc = new Builder().issuerIdNumber("123456").build();
-        assertEquals("123456", cc.getIssuerIdNumber());
+        assertEquals("123456", cc.issuerIdNumber());
 
         cc = new Builder().issuerIdNumber("12345678").build();
-        assertEquals("12345678", cc.getIssuerIdNumber());
+        assertEquals("12345678", cc.issuerIdNumber());
     }
 
     @Test
@@ -47,10 +47,10 @@ public class CreditCardTest {
     @Test
     public void testLastDigits() {
         CreditCard cc = new Builder().lastDigits("1234").build();
-        assertEquals("1234", cc.getLastDigits());
+        assertEquals("1234", cc.lastDigits());
 
         cc = new Builder().lastDigits("12").build();
-        assertEquals("12", cc.getLastDigits());
+        assertEquals("12", cc.lastDigits());
     }
 
     @Test
@@ -81,27 +81,27 @@ public class CreditCardTest {
     @Test
     public void testBankName() {
         CreditCard cc = new Builder().bankName("Bank").build();
-        assertEquals("Bank", cc.getBankName());
+        assertEquals("Bank", cc.bankName());
     }
 
     @Test
     public void testBankPhoneCountryCode() {
         CreditCard cc = new Builder().bankPhoneCountryCode("1").build();
-        assertEquals("1", cc.getBankPhoneCountryCode());
+        assertEquals("1", cc.bankPhoneCountryCode());
     }
 
     @Test
     public void testBankPhoneNumber() {
         String phone = "231-323-3123";
         CreditCard cc = new Builder().bankPhoneNumber(phone).build();
-        assertEquals(phone, cc.getBankPhoneNumber());
+        assertEquals(phone, cc.bankPhoneNumber());
     }
 
     @Test
     public void testCountry() {
         String country = "CA";
         CreditCard cc = new Builder().country(country).build();
-        assertEquals(country, cc.getCountry());
+        assertEquals(country, cc.country());
     }
 
     @ParameterizedTest
@@ -116,13 +116,13 @@ public class CreditCardTest {
     @Test
     public void testAvsResult() {
         CreditCard cc = new Builder().avsResult('Y').build();
-        assertEquals(Character.valueOf('Y'), cc.getAvsResult());
+        assertEquals(Character.valueOf('Y'), cc.avsResult());
     }
 
     @Test
     public void testCvvResult() {
         CreditCard cc = new Builder().cvvResult('N').build();
-        assertEquals(Character.valueOf('N'), cc.getCvvResult());
+        assertEquals(Character.valueOf('N'), cc.cvvResult());
     }
 
     @ParameterizedTest
@@ -146,12 +146,12 @@ public class CreditCardTest {
     })
     public void testValidToken(String token) {
         CreditCard cc = new Builder().token(token).build();
-        assertEquals(token, cc.getToken());
+        assertEquals(token, cc.token());
     }
 
     @Test
     public void testWas3dSecureSuccessful() {
         CreditCard cc = new Builder().was3dSecureSuccessful(true).build();
-        assertTrue(cc.getWas3dSecureSuccessful());
+        assertTrue(cc.was3dSecureSuccessful());
     }
 }