+++ src/test/java/com/maxmind/minfraud/request/DeviceTest.java
@@ -19,46 +19,46 @@ public class DeviceTest {
     @Test
     public void testConstructorWithoutIP() {
         Device device = new Builder().build();
-        assertNull(device.getIpAddress());
+        assertNull(device.ipAddress());
     }
 
     @Test
     public void testIpAddressThroughConstructor() {
         Device device = new Builder(ip).build();
-        assertEquals(ip, device.getIpAddress());
+        assertEquals(ip, device.ipAddress());
     }
 
     @Test
     public void testIpAddress() {
         Device device = new Builder().ipAddress(ip).build();
-        assertEquals(ip, device.getIpAddress());
+        assertEquals(ip, device.ipAddress());
     }
 
     @Test
     public void testUserAgent() {
         String ua = "Mozila 5";
         Device device = new Builder(ip).userAgent(ua).build();
-        assertEquals(ua, device.getUserAgent());
+        assertEquals(ua, device.userAgent());
     }
 
     @Test
     public void testAcceptLanguage() {
         String al = "en-US";
         Device device = new Builder(ip).acceptLanguage(al).build();
-        assertEquals(al, device.getAcceptLanguage());
+        assertEquals(al, device.acceptLanguage());
     }
 
     @Test
     public void testSessionAge() {
         Double hour = 3600d;
         Device device = new Builder(ip).sessionAge(hour).build();
-        assertEquals(hour, device.getSessionAge());
+        assertEquals(hour, device.sessionAge());
     }
 
     @Test
     public void testSessionId() {
         String id = "foobar";
         Device device = new Builder(ip).sessionId(id).build();
-        assertEquals(id, device.getSessionId());
+        assertEquals(id, device.sessionId());
     }
 }