+++ src/test/java/com/maxmind/minfraud/request/EmailTest.java
@@ -21,15 +21,15 @@ public class EmailTest {
     @Test
     public void testAddress() {
         Email email = new Builder().address("test@test.org").build();
-        assertEquals("test@test.org", email.getAddress(), "raw email");
-        assertEquals("test.org", email.getDomain(), "domain set from email");
+        assertEquals("test@test.org", email.address(), "raw email");
+        assertEquals("test.org", email.domain(), "domain set from email");
     }
 
     @Test
     public void testMultipleAtAddress() {
         Email email = new Builder().address("\"test@test\"@test.org").build();
-        assertEquals("\"test@test\"@test.org", email.getAddress(), "raw email");
-        assertEquals("test.org", email.getDomain(), "domain set from email");
+        assertEquals("\"test@test\"@test.org", email.address(), "raw email");
+        assertEquals("test.org", email.domain(), "domain set from email");
     }
 
     @Test
@@ -43,8 +43,8 @@ public class EmailTest {
 
         for (String address : addresses.keySet()) {
             Email email = new Builder(false).address(address).build();
-            assertEquals(address, email.getAddress(), "raw email");
-            assertEquals(addresses.get(address), email.getDomain(), "domain set from email");
+            assertEquals(address, email.address(), "raw email");
+            assertEquals(addresses.get(address), email.domain(), "domain set from email");
         }
     }
 
@@ -53,10 +53,10 @@ public class EmailTest {
         Email email = new Builder().address("test@test.org").hashAddress().build();
         assertEquals(
             "476869598e748d958e819c180af31982",
-            email.getAddress(),
+            email.address(),
             "MD5 generated from email"
         );
-        assertEquals("test.org", email.getDomain(), "domain set from email");
+        assertEquals("test.org", email.domain(), "domain set from email");
     }
 
     @Test
@@ -64,16 +64,16 @@ public class EmailTest {
         Email email = new Builder().address("test@test.org").hashAddress().hashAddress().build();
         assertEquals(
             "476869598e748d958e819c180af31982",
-            email.getAddress(),
+            email.address(),
             "MD5 generated from email"
         );
-        assertEquals("test.org", email.getDomain(), "domain set from email");
+        assertEquals("test.org", email.domain(), "domain set from email");
     }
 
     @Test
     public void testHashAddressWithoutAddress() {
         Email email = new Builder().domain("test.org").hashAddress().build();
-        assertEquals("test.org", email.getDomain(), "domain is set");
+        assertEquals("test.org", email.domain(), "domain is set");
     }
 
     @Test
@@ -81,7 +81,7 @@ public class EmailTest {
         Email email = new Builder().address("TEST@TEST.org").hashAddress().build();
         assertEquals(
             "476869598e748d958e819c180af31982",
-            email.getAddress(),
+            email.address(),
             "MD5 generated from lowercased email"
         );
     }
@@ -89,10 +89,10 @@ public class EmailTest {
     @Test
     public void testGetAddressWithoutSettingIt() {
         Email email = new Builder().domain("test.org").hashAddress().build();
-        assertNull(email.getAddress(), "null address if none set");
+        assertNull(email.address(), "null address if none set");
 
         Email email2 = new Builder().domain("test.org").hashAddress().build();
-        assertNull(email2.getAddress(), "null address if none set");
+        assertNull(email2.address(), "null address if none set");
     }
 
     @Test
@@ -100,142 +100,142 @@ public class EmailTest {
         Email e;
 
         e = new Builder().address("test@maxmind.com").hashAddress().build();
-        assertEquals("977577b140bfb7c516e4746204fbdb01", e.getAddress(), "MD5");
-        assertEquals("maxmind.com", e.getDomain(), "domain");
+        assertEquals("977577b140bfb7c516e4746204fbdb01", e.address(), "MD5");
+        assertEquals("maxmind.com", e.domain(), "domain");
 
         e = new Builder().address("Test@maxmind.com").hashAddress().build();
-        assertEquals("977577b140bfb7c516e4746204fbdb01", e.getAddress(), "MD5");
-        assertEquals("maxmind.com", e.getDomain(), "domain");
+        assertEquals("977577b140bfb7c516e4746204fbdb01", e.address(), "MD5");
+        assertEquals("maxmind.com", e.domain(), "domain");
 
         e = new Builder(false).address("  Test@maxmind.com").hashAddress().build();
-        assertEquals("977577b140bfb7c516e4746204fbdb01", e.getAddress(), "MD5");
-        assertEquals("maxmind.com", e.getDomain(), "domain");
+        assertEquals("977577b140bfb7c516e4746204fbdb01", e.address(), "MD5");
+        assertEquals("maxmind.com", e.domain(), "domain");
 
         e = new Builder().address("Test+alias@maxmind.com").hashAddress().build();
-        assertEquals("977577b140bfb7c516e4746204fbdb01", e.getAddress(), "MD5");
-        assertEquals("maxmind.com", e.getDomain(), "domain");
+        assertEquals("977577b140bfb7c516e4746204fbdb01", e.address(), "MD5");
+        assertEquals("maxmind.com", e.domain(), "domain");
 
         e = new Builder().address("Test+007+008@maxmind.com").hashAddress().build();
-        assertEquals("977577b140bfb7c516e4746204fbdb01", e.getAddress(), "MD5");
-        assertEquals("maxmind.com", e.getDomain(), "domain");
+        assertEquals("977577b140bfb7c516e4746204fbdb01", e.address(), "MD5");
+        assertEquals("maxmind.com", e.domain(), "domain");
 
         e = new Builder().address("Test+@maxmind.com").hashAddress().build();
-        assertEquals("977577b140bfb7c516e4746204fbdb01", e.getAddress(), "MD5");
-        assertEquals("maxmind.com", e.getDomain(), "domain");
+        assertEquals("977577b140bfb7c516e4746204fbdb01", e.address(), "MD5");
+        assertEquals("maxmind.com", e.domain(), "domain");
 
         e = new Builder(false).address("Test@maxmind.com.").hashAddress().build();
-        assertEquals("977577b140bfb7c516e4746204fbdb01", e.getAddress(), "MD5");
-        assertEquals("maxmind.com.", e.getDomain(), "domain");
+        assertEquals("977577b140bfb7c516e4746204fbdb01", e.address(), "MD5");
+        assertEquals("maxmind.com.", e.domain(), "domain");
 
         e = new Builder().address("+@maxmind.com").hashAddress().build();
-        assertEquals("aa57884e48f0dda9fc6f4cb2bffb1dd2", e.getAddress(), "MD5");
-        assertEquals("maxmind.com", e.getDomain(), "domain");
+        assertEquals("aa57884e48f0dda9fc6f4cb2bffb1dd2", e.address(), "MD5");
+        assertEquals("maxmind.com", e.domain(), "domain");
 
         e = new Builder(false).address("Test@ maxmind.com").hashAddress().build();
-        assertEquals("977577b140bfb7c516e4746204fbdb01", e.getAddress(), "MD5");
-        assertEquals(" maxmind.com", e.getDomain(), "domain");
+        assertEquals("977577b140bfb7c516e4746204fbdb01", e.address(), "MD5");
+        assertEquals(" maxmind.com", e.domain(), "domain");
 
         e = new Builder().address("Test+foo@yahoo.com").hashAddress().build();
-        assertEquals("a5f830c699fd71ad653aa59fa688c6d9", e.getAddress(), "MD5");
-        assertEquals("yahoo.com", e.getDomain(), "domain");
+        assertEquals("a5f830c699fd71ad653aa59fa688c6d9", e.address(), "MD5");
+        assertEquals("yahoo.com", e.domain(), "domain");
 
         e = new Builder().address("Test-foo@yahoo.com").hashAddress().build();
-        assertEquals("88e478531ab3bc303f1b5da82c2e9bbb", e.getAddress(), "MD5");
-        assertEquals("yahoo.com", e.getDomain(), "domain");
+        assertEquals("88e478531ab3bc303f1b5da82c2e9bbb", e.address(), "MD5");
+        assertEquals("yahoo.com", e.domain(), "domain");
 
         e = new Builder().address("Test-foo-foo2@yahoo.com").hashAddress().build();
-        assertEquals("88e478531ab3bc303f1b5da82c2e9bbb", e.getAddress(), "MD5");
-        assertEquals("yahoo.com", e.getDomain(), "domain");
+        assertEquals("88e478531ab3bc303f1b5da82c2e9bbb", e.address(), "MD5");
+        assertEquals("yahoo.com", e.domain(), "domain");
 
         e = new Builder().address("Test-foo@gmail.com").hashAddress().build();
-        assertEquals("6f3ff986fa5e830dbbf08a942777a17c", e.getAddress(), "MD5");
-        assertEquals("gmail.com", e.getDomain(), "domain");
+        assertEquals("6f3ff986fa5e830dbbf08a942777a17c", e.address(), "MD5");
+        assertEquals("gmail.com", e.domain(), "domain");
 
         e = new Builder().address("test@gmail.com").hashAddress().build();
-        assertEquals("1aedb8d9dc4751e229a335e371db8058", e.getAddress(), "MD5");
-        assertEquals("gmail.com", e.getDomain(), "domain");
+        assertEquals("1aedb8d9dc4751e229a335e371db8058", e.address(), "MD5");
+        assertEquals("gmail.com", e.domain(), "domain");
 
         e = new Builder().address("test@gamil.com").hashAddress().build();
-        assertEquals("1aedb8d9dc4751e229a335e371db8058", e.getAddress(), "MD5");
-        assertEquals("gamil.com", e.getDomain(), "domain");
+        assertEquals("1aedb8d9dc4751e229a335e371db8058", e.address(), "MD5");
+        assertEquals("gamil.com", e.domain(), "domain");
 
         e = new Builder().address("test@bücher.com").hashAddress().build();
-        assertEquals("24948acabac551360cd510d5e5e2b464", e.getAddress(), "MD5");
-        assertEquals("bücher.com", e.getDomain(), "domain");
+        assertEquals("24948acabac551360cd510d5e5e2b464", e.address(), "MD5");
+        assertEquals("bücher.com", e.domain(), "domain");
 
         e = new Builder().address("Test+alias@Bücher.com").hashAddress().build();
-        assertEquals("24948acabac551360cd510d5e5e2b464", e.getAddress(), "MD5");
-        assertEquals("Bücher.com", e.getDomain(), "domain");
+        assertEquals("24948acabac551360cd510d5e5e2b464", e.address(), "MD5");
+        assertEquals("Bücher.com", e.domain(), "domain");
 
         e = new Builder(false).address("test").hashAddress().build();
-        assertEquals("098f6bcd4621d373cade4e832627b4f6", e.getAddress(), "MD5");
-        assertNull(e.getDomain(), "domain");
+        assertEquals("098f6bcd4621d373cade4e832627b4f6", e.address(), "MD5");
+        assertNull(e.domain(), "domain");
 
         e = new Builder(false).address("test@").hashAddress().build();
-        assertEquals("246a848af2f8394e3adbc738dbe43720", e.getAddress(), "MD5");
-        assertNull(e.getDomain(), "domain");
+        assertEquals("246a848af2f8394e3adbc738dbe43720", e.address(), "MD5");
+        assertNull(e.domain(), "domain");
 
         e = new Builder(false).address("test@.").hashAddress().build();
-        assertEquals("246a848af2f8394e3adbc738dbe43720", e.getAddress(), "MD5");
-        assertEquals(".", e.getDomain(), "domain");
+        assertEquals("246a848af2f8394e3adbc738dbe43720", e.address(), "MD5");
+        assertEquals(".", e.domain(), "domain");
 
         e = new Builder(false).address("foo@googlemail.com").hashAddress().build();
-        assertEquals(toMD5("foo@gmail.com"), e.getAddress(), "MD5");
-        assertEquals("googlemail.com", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@gmail.com"), e.address(), "MD5");
+        assertEquals("googlemail.com", e.domain(), "domain");
 
         e = new Builder(false).address("foo.bar@gmail.com").hashAddress().build();
-        assertEquals(toMD5("foobar@gmail.com"), e.getAddress(), "MD5");
-        assertEquals("gmail.com", e.getDomain(), "domain");
+        assertEquals(toMD5("foobar@gmail.com"), e.address(), "MD5");
+        assertEquals("gmail.com", e.domain(), "domain");
 
         e = new Builder(false).address("alias@user.fastmail.com").hashAddress().build();
-        assertEquals(toMD5("user@fastmail.com"), e.getAddress(), "MD5");
-        assertEquals("user.fastmail.com", e.getDomain(), "domain");
+        assertEquals(toMD5("user@fastmail.com"), e.address(), "MD5");
+        assertEquals("user.fastmail.com", e.domain(), "domain");
 
         e = new Builder(false).address("foo@bar.example.com").hashAddress().build();
-        assertEquals(toMD5("foo@bar.example.com"), e.getAddress(), "MD5");
-        assertEquals("bar.example.com", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@bar.example.com"), e.address(), "MD5");
+        assertEquals("bar.example.com", e.domain(), "domain");
 
         e = new Builder(false).address("foo-bar@ymail.com").hashAddress().build();
-        assertEquals(toMD5("foo@ymail.com"), e.getAddress(), "MD5");
-        assertEquals("ymail.com", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@ymail.com"), e.address(), "MD5");
+        assertEquals("ymail.com", e.domain(), "domain");
 
         e = new Builder(false).address("foo@example.com.com").hashAddress().build();
-        assertEquals(toMD5("foo@example.com"), e.getAddress(), "MD5");
-        assertEquals("example.com.com", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@example.com"), e.address(), "MD5");
+        assertEquals("example.com.com", e.domain(), "domain");
 
         e = new Builder(false).address("foo@example.comfoo").hashAddress().build();
-        assertEquals(toMD5("foo@example.comfoo"), e.getAddress(), "MD5");
-        assertEquals("example.comfoo", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@example.comfoo"), e.address(), "MD5");
+        assertEquals("example.comfoo", e.domain(), "domain");
 
         e = new Builder(false).address("foo@example.cam").hashAddress().build();
-        assertEquals(toMD5("foo@example.cam"), e.getAddress(), "MD5");
-        assertEquals("example.cam", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@example.cam"), e.address(), "MD5");
+        assertEquals("example.cam", e.domain(), "domain");
 
         e = new Builder(false).address("foo@10000gmail.com").hashAddress().build();
-        assertEquals(toMD5("foo@gmail.com"), e.getAddress(), "MD5");
-        assertEquals("10000gmail.com", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@gmail.com"), e.address(), "MD5");
+        assertEquals("10000gmail.com", e.domain(), "domain");
 
         e = new Builder(false).address("foo@example.comcom").hashAddress().build();
-        assertEquals(toMD5("foo@example.com"), e.getAddress(), "MD5");
-        assertEquals("example.comcom", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@example.com"), e.address(), "MD5");
+        assertEquals("example.comcom", e.domain(), "domain");
 
         e = new Builder(false).address("foo@example.com.").hashAddress().build();
-        assertEquals(toMD5("foo@example.com"), e.getAddress(), "MD5");
-        assertEquals("example.com.", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@example.com"), e.address(), "MD5");
+        assertEquals("example.com.", e.domain(), "domain");
 
         e = new Builder(false).address("foo@example.com...").hashAddress().build();
-        assertEquals(toMD5("foo@example.com"), e.getAddress(), "MD5");
-        assertEquals("example.com...", e.getDomain(), "domain");
+        assertEquals(toMD5("foo@example.com"), e.address(), "MD5");
+        assertEquals("example.com...", e.domain(), "domain");
 
         e = new Builder().address("example@bu\u0308cher.com").hashAddress().build();
-        assertEquals("2b21bc76dab3c8b1622837c1d698936c", e.getAddress(), "MD5");
+        assertEquals("2b21bc76dab3c8b1622837c1d698936c", e.address(), "MD5");
         e = new Builder().address("example@b\u00FCcher.com").hashAddress().build();
-        assertEquals("2b21bc76dab3c8b1622837c1d698936c", e.getAddress(), "MD5");
+        assertEquals("2b21bc76dab3c8b1622837c1d698936c", e.address(), "MD5");
 
         e = new Builder().address("bu\u0308cher@example.com").hashAddress().build();
-        assertEquals("53550c712b146287a2d0dd30e5ed6f4b", e.getAddress(), "MD5");
+        assertEquals("53550c712b146287a2d0dd30e5ed6f4b", e.address(), "MD5");
         e = new Builder().address("b\u00FCcher@example.com").hashAddress().build();
-        assertEquals("53550c712b146287a2d0dd30e5ed6f4b", e.getAddress(), "MD5");
+        assertEquals("53550c712b146287a2d0dd30e5ed6f4b", e.address(), "MD5");
     }
 
     private String toMD5(String s) throws NoSuchAlgorithmException {
@@ -272,14 +272,14 @@ public class EmailTest {
     public void testDomain() {
         String domain = "domain.com";
         Email email = new Builder().domain(domain).build();
-        assertEquals(domain, email.getDomain());
+        assertEquals(domain, email.domain());
     }
 
     @Test
     public void testDomainWithoutValidation() {
         String domain = "bad domain @!";
         Email email = new Builder(false).domain(domain).build();
-        assertEquals(domain, email.getDomain());
+        assertEquals(domain, email.domain());
     }
 
     @ParameterizedTest(name = "Run #{index}: domain = \"{0}\"")