+++ src/test/java/com/maxmind/minfraud/request/OrderTest.java
@@ -14,19 +14,19 @@ public class OrderTest {
     @Test
     public void testDoubleAmount() {
         Order order = new Builder().amount(1.1).build();
-        assertEquals(BigDecimal.valueOf(1.1), order.getAmount());
+        assertEquals(BigDecimal.valueOf(1.1), order.amount());
     }
 
     @Test
     public void testAmount() {
         Order order = new Builder().amount(BigDecimal.valueOf(1.1)).build();
-        assertEquals(BigDecimal.valueOf(1.1), order.getAmount());
+        assertEquals(BigDecimal.valueOf(1.1), order.amount());
     }
 
     @Test
     public void testCurrency() {
         Order order = new Builder().currency("USD").build();
-        assertEquals("USD", order.getCurrency());
+        assertEquals("USD", order.currency());
     }
 
     @Test
@@ -65,26 +65,26 @@ public class OrderTest {
     @Test
     public void testDiscountCode() {
         Order order = new Builder().discountCode("dsc").build();
-        assertEquals("dsc", order.getDiscountCode());
+        assertEquals("dsc", order.discountCode());
     }
 
     @Test
     public void testAffiliateId() {
         Order order = new Builder().affiliateId("af").build();
-        assertEquals("af", order.getAffiliateId());
+        assertEquals("af", order.affiliateId());
     }
 
     @Test
     public void testSubaffiliateId() {
         Order order = new Builder().subaffiliateId("saf").build();
-        assertEquals("saf", order.getSubaffiliateId());
+        assertEquals("saf", order.subaffiliateId());
     }
 
     @Test
     public void testReferrerUri() throws Exception {
         URI uri = new URI("http://www.mm.com/");
         Order order = new Builder().referrerUri(uri).build();
-        assertEquals(uri, order.getReferrerUri());
+        assertEquals(uri, order.referrerUri());
     }
 
     @Test