+++ src/test/java/com/maxmind/minfraud/request/ShippingTest.java
@@ -15,6 +15,6 @@ public class ShippingTest extends AbstractLocationTest {
     @Test
     public void testDeliverySpeed() {
         Shipping loc = this.builder().deliverySpeed(DeliverySpeed.EXPEDITED).build();
-        assertEquals(DeliverySpeed.EXPEDITED, loc.getDeliverySpeed());
+        assertEquals(DeliverySpeed.EXPEDITED, loc.deliverySpeed());
     }
 }