+++ src/test/java/com/maxmind/minfraud/request/TransactionTest.java
@@ -16,41 +16,41 @@ public class TransactionTest {
     @Test
     public void testConstructorWithoutDevice() {
         Transaction request = new Transaction.Builder().build();
-        assertNull(request.getDevice());
+        assertNull(request.device());
     }
 
     @Test
     public void testAccount() throws Exception {
         Transaction request =
             this.builder().account(new Account.Builder().userId("1").build()).build();
-        assertEquals("1", request.getAccount().getUserId());
+        assertEquals("1", request.account().userId());
     }
 
     @Test
     public void testBilling() throws Exception {
         Transaction request =
             this.builder().billing(new Billing.Builder().address("add").build()).build();
-        assertEquals("add", request.getBilling().getAddress());
+        assertEquals("add", request.billing().address());
     }
 
     @Test
     public void testCreditCard() throws Exception {
         Transaction request =
             this.builder().creditCard(new CreditCard.Builder().bankName("name").build()).build();
-        assertEquals("name", request.getCreditCard().getBankName());
+        assertEquals("name", request.creditCard().bankName());
     }
 
     @Test
     public void testCustomInputs() throws Exception {
         Transaction request = this.builder().customInputs(
             new CustomInputs.Builder().put("key", "value").build()).build();
-        assertEquals("value", request.getCustomInputs().getInputs().get("key"));
+        assertEquals("value", request.customInputs().inputs().get("key"));
     }
 
     @Test
     public void testDevice() throws Exception {
         Transaction request = this.builder().build();
-        assertEquals(InetAddress.getByName("152.216.7.110"), request.getDevice().getIpAddress());
+        assertEquals(InetAddress.getByName("152.216.7.110"), request.device().ipAddress());
     }
 
     @Test
@@ -63,41 +63,41 @@ public class TransactionTest {
             .device(device)
             .build();
 
-        assertEquals(ip, request.getDevice().getIpAddress());
+        assertEquals(ip, request.device().ipAddress());
     }
 
     @Test
     public void testEmail() throws Exception {
         Transaction request =
             this.builder().email(new Email.Builder().domain("test.com").build()).build();
-        assertEquals("test.com", request.getEmail().getDomain());
+        assertEquals("test.com", request.email().domain());
     }
 
     @Test
     public void testEvent() throws Exception {
         Transaction request = this.builder().event(new Event.Builder().shopId("1").build()).build();
-        assertEquals("1", request.getEvent().getShopId());
+        assertEquals("1", request.event().shopId());
     }
 
     @Test
     public void testOrder() throws Exception {
         Transaction request =
             this.builder().order(new Order.Builder().affiliateId("af1").build()).build();
-        assertEquals("af1", request.getOrder().getAffiliateId());
+        assertEquals("af1", request.order().affiliateId());
     }
 
     @Test
     public void testPayment() throws Exception {
         Transaction request =
             this.builder().payment(new Payment.Builder().declineCode("d").build()).build();
-        assertEquals("d", request.getPayment().getDeclineCode());
+        assertEquals("d", request.payment().declineCode());
     }
 
     @Test
     public void testShipping() throws Exception {
         Transaction request =
             this.builder().shipping(new Shipping.Builder().lastName("l").build()).build();
-        assertEquals("l", request.getShipping().getLastName());
+        assertEquals("l", request.shipping().lastName());
     }
 
     @Test
@@ -105,6 +105,6 @@ public class TransactionTest {
         Transaction request =
             this.builder().addShoppingCartItem(new ShoppingCartItem.Builder().itemId("1").build())
                 .build();
-        assertEquals("1", request.getShoppingCart().get(0).getItemId());
+        assertEquals("1", request.shoppingCart().get(0).itemId());
     }
 }